fixed links in app so they access components
