+++ src/components/Search.js
@@ -1,0 +1,20 @@
+import React, { Component } from 'react';
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link,
+  useRouteMatch,
+  useParams
+} from 'react-router-dom';
+import PropTypes from 'prop-types';
+
+const Search = (props) => {
+  return(
+    <div>
+
+    </div>
+  )
+
+}
+export default Search
